🐛 fix(web): resolver pendencias do quality gate da space page
